requirements.txt
semver
typed-argument-parser
numpy
dataclasses-json
tqdm
paddlepaddle
paddlenlp
loguru
visualdl